Move System.ValueType.s_seed into System.Runtime.CompilerServices.RuntimeHelpers
PR https://github.com/dotnet/runtime/pull/69723 added a 's_seed' field to
System.ValueType. This causes the debugger to show all structs as having
static members. This PR moves the field and the function that used it into RuntimeHelpers.
