+++ src/coreclr/System.Private.CoreLib/src/System/ArgIterator.cs
@@ -123,7 +123,7 @@ namespace System
 
         public override int GetHashCode()
         {
-            return ValueType.GetHashCodeOfPtr(ArgCookie);
+            return RuntimeHelpers.GetHashCodeOfPtr(ArgCookie);
         }
 
         // Inherited from object
